link to webrecorder
Clicking on URLs and domains in the reports takes you over to Webrecorder where
you can view that URL and capture it into an anonymous collection.

This isn't particularly useful since there is no continuity between the user on
dnflow and webrecorder. But it should at least help frame some of the discussion
about a tighter coordination between docnow and webrecorder in order to allow
researchers and archivists to identify material in need of archiving.

Fixes #11