basic auth
This is a temporary hack to get the PUT to /job working in our demo
environment where the Flask application is running behind Nginx with HTTP
Basic Authentication. Basic Auth is only being used temporarily so we can share
the application with a group of people at our STL meeting.

The use of Basic Authentication is dependent on two variables in `dnflow.cfg`:

* HTTP_BASICAUTH_USER
* HTTP_BASICAUTH_PASS

If these are not defined the luigi tasks will fall back to using no auth on
the PUT requests to /job.

Eventually we'll need to figure out a secure approach to doing database
updates from Luigi tasks, so not anyone on the Internet can update jobs. This
is why we do prototypes :)

refs #24