inactive search links
This commit includes a bit of logic to prevent searches on the homepage from
being linked when the workflow is not complete. This will prevent users from
clicking on the search when data is incomplete, and the report is in an odd
state.

This commit also includes a modification to get_block_size to allow for a
dampening value to be used when some jobs run slower (FetchMedia & MatchMedia).