publishing & deleting
Added the ability to publish, unpublish and delete datasets. A user must
own the dataset in order to perform these actions. Before a dataset is
published it cannot be viewed by anyone except the creator. Once it is
published it is viewable by the world.

An option on the homepage allows a user to see their datasets alone or
their datasets along with all published datasets. A non-logged in user
can only see published datasets.

Fixes #37
Fixes #38
Fixes #40